+++ ui/src/js/simpleweb/boundary/pages/commander/dialogs/index.ts
@@ -1,5 +1,6 @@
 // Dialog Components Barrel Export
 export * from './CommandDialog.js'
+export * from './ContextMenuDialog.js'
 export * from './DeleteDialog.js'
 export * from './DriveSelectorDialog.js'
 export * from './HelpDialog.js'