+++ version/changelog.md
@@ -6,6 +6,8 @@
 
 - in drive dialog support keyboard navigation
 - memorize directories and enable backward history navigation
+- support nested zips
+- add context menu using SHIFT+F10
 
 ### v1.5.0 (31.01.2026)